Adapted associated controllers with the new role and whisky properties. Restricted methods for certain roles only.

+++ TheWhiskyRealm/Controllers/AwardController.cs
@@ -1,8 +1,10 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
 using System.Security.Claims;
 using TheWhiskyRealm.Core.Contracts;
 using TheWhiskyRealm.Core.Models.Award;
 using TheWhiskyRealm.Infrastructure.Data.Enums;
+using TheWhiskyRealm.Infrastructure.Data.Models;
 
 namespace TheWhiskyRealm.Controllers;
 
@@ -17,14 +19,11 @@ public class AwardController : BaseController
         this.whiskyService = whiskyService;
     }
 
+    [Authorize(Roles = "Administrator, WhiskyExpert")]
     [HttpGet]
     public async Task<IActionResult> Edit(int id)
     {
         var userId = User.Id(); //TODO Remove all unnecessary validations like this
-        if (userId == null)
-        {
-            return RedirectToPage("/Account/Login");
-        }
 
         if (await awardService.AwardExistAsync(id) == false)
         {
@@ -33,18 +32,28 @@ public class AwardController : BaseController
 
         var model = await awardService.GetAwardByIdAsync(id);
 
+        var publisherId = await whiskyService.GetWhiskyPublisherAsync(model.WhiskyId);
+
+        if (User.IsInRole("WhiskyExpert") && publisherId != User.Id())
+        {
+            return Unauthorized();
+        }
+
         model.MedalTypeOptions = Enum.GetNames(typeof(MedalType));
 
         return View(model);
     }
 
+    [Authorize(Roles = "Administrator, WhiskyExpert")]
     [HttpPost]
     public async Task<IActionResult> Edit(AwardViewModel model)
     {
         var userId = User.Id();
-        if (userId == null)
+        var publisherId = await whiskyService.GetWhiskyPublisherAsync(model.WhiskyId);
+
+        if (User.IsInRole("WhiskyExpert") && publisherId != User.Id())
         {
-            return RedirectToPage("/Account/Login");
+            return Unauthorized();
         }
 
         if (await awardService.AwardExistAsync(model.Id) == false)
@@ -70,14 +79,11 @@ public class AwardController : BaseController
         return RedirectToAction("Details", "Whisky", new { id = model.WhiskyId });
     }
 
+    [Authorize(Roles = "Administrator, WhiskyExpert")]
     [HttpGet]
     public async Task<IActionResult> Delete(int id)
     {
         var userId = User.Id();
-        if (userId == null)
-        {
-            return RedirectToPage("/Account/Login");
-        }
 
         var award = await awardService.GetAwardByIdAsync(id);
         if (award == null)
@@ -85,17 +91,21 @@ public class AwardController : BaseController
             return NotFound();
         }
 
+        var publisherId = await whiskyService.GetWhiskyPublisherAsync(award.WhiskyId);
+
+        if (User.IsInRole("WhiskyExpert") && publisherId != User.Id())
+        {
+            return Unauthorized();
+        }
+
         return View(award);
     }
 
+    [Authorize(Roles = "Administrator, WhiskyExpert")]
     [HttpPost]
     public async Task<IActionResult> Delete(AwardViewModel model)
     {
         var userId = User.Id();
-        if (userId == null)
-        {
-            return RedirectToPage("/Account/Login");
-        }
 
         var award = await awardService.GetAwardByIdAsync(model.Id);
         if (award == null)
@@ -103,11 +113,19 @@ public class AwardController : BaseController
             return NotFound();
         }
 
+        var publisherId = await whiskyService.GetWhiskyPublisherAsync(award.WhiskyId);
+
+        if (User.IsInRole("WhiskyExpert") && publisherId != User.Id())
+        {
+            return Unauthorized();
+        }
+
         await awardService.DeleteAwardAsync(model.Id);
 
         return RedirectToAction("Details", "Whisky", new { id = award.WhiskyId });
     }
 
+    [Authorize(Roles = "Administrator, WhiskyExpert")]
     [HttpGet]
     public async Task<IActionResult> Add(int id)
     {
@@ -121,6 +139,14 @@ public class AwardController : BaseController
         {
             return NotFound();
         }
+
+        var publisherId = await whiskyService.GetWhiskyPublisherAsync(id);
+
+        if (User.IsInRole("WhiskyExpert") && publisherId != User.Id())
+        {
+            return Unauthorized();
+        }
+
         var model = new AwardAddModel()
         {
             MedalTypeOptions = Enum.GetNames(typeof(MedalType)),
@@ -131,6 +157,7 @@ public class AwardController : BaseController
         return View(model);
     }
 
+    [Authorize(Roles = "Administrator, WhiskyExpert")]
     [HttpPost]
     public async Task<IActionResult> Add(AwardAddModel model)
     {
@@ -143,6 +170,13 @@ public class AwardController : BaseController
         if (await whiskyService.WhiskyExistAsync(model.WhiskyId) == false)
         {
             return NotFound();
+        }
+
+        var publisherId = await whiskyService.GetWhiskyPublisherAsync(model.WhiskyId);
+
+        if (User.IsInRole("WhiskyExpert") && publisherId != User.Id())
+        {
+            return Unauthorized();
         }
 
         if (model.MedalType != "Gold" && model.MedalType != "Silver" && model.MedalType != "Bronze")
